Update: Move Projects commponent and all its child components
