+++ src/components/HomePage/Modal/ModalMobile/ModalMobile.tsx
@@ -4,7 +4,7 @@ import { HiUserGroup } from 'react-icons/hi'
 import { IoIosArrowBack, IoIosArrowForward } from 'react-icons/io'
 import { MdClose } from 'react-icons/md'
 
-import { IProject } from '../../../../../common/interfaces'
+import { IProject } from '../../../../common/interfaces'
 import ProjectInfo from '../../ProjectInfo'
 import ModalTeam from '../ModalTeam'
 import './styles.css'
