+++ src/components/HomePage/ProjectInfo/ProjectInfo.tsx
