+++ src/components/HomePage/ProjectInfo/index.ts
