+++ src/components/HomePage/ProjectInfo/styles.css
