+++ src/components/HomePage/ScrollAnimation/ScrollAnimation.tsx
