+++ src/components/HomePage/ScrollAnimation/index.ts
