+++ src/components/HomePage/Testimonials/index.ts
