+++ src/pages/Home/Home.tsx
@@ -1,13 +1,13 @@
 import React, { FunctionComponent, ReactElement } from 'react'
 
 import Footer from '@/components/HomePage/Footer'
+import Projects from '@/components/HomePage/Projects'
+import Testimonials from '@/components/HomePage/Testimonials'
 import WelcomeScreen from '@/components/HomePage/WelcomeScreen'
 import data from '@/data'
 
 import { AppContext } from '@/context/useAppContext'
 
-import Projects from './Projects'
-import Testimonials from './Testimonials'
 import './styles.css'
 
 const Home: FunctionComponent = (): ReactElement => {
